Rename geographycale_zone module
